Changed main arg from unsigned 8 bit integer to standard integer in accordance to C standard

+++ scaled_to_unscaled.cpp
@@ -9,7 +9,7 @@ void convertToScaled(const std::string& sampleName, const std::vector<float>& co
     std::cout<<"\n"; 
 }
 
-int main(uint8_t argc, char*argv[]) {
+int main(int argc, char*argv[]) {
     if(argc<2) error_handle(1); 
     std::string filePath = argv[1]; 
     BatchProcessor convertUnscaledtoScaled(filePath, convertToScaled);
